#14 introducing a new telemetry struct
