--- pkg/telemetry/config.go
@@ -1,7 +1,0 @@
-package telemetry
-
-import "go.uber.org/zap"
-
-type TelemetryConfig struct {
-	LogConfig *zap.Config `json:"logs" yaml:"logs"`
-}
